Use composer installed PHPUnit in build
This is to avoid getting a crapton of warnings when using recent PHPUnit
versions. Fun fact: we cannot avoid those warnings without becomming
unsable to run the build on PHPUnit older than 5.6 due to recent PHPUnit
requiring 5.6 or later, and it deprecating things for which the alternative
was introduced in those same versions. #MAW