Support mediawiki-codesniffer@45.0.0, PHP 8.4 readiness
This repository is currently using version 34 of the Mediawiki
coding style phpcs rules. It should be upgraded to the latest version.

Besides being simply different, the newer versions of the rules
introduce, for example,
MediaWiki.Usage.NullableType.ExplicitNullableTypes, which enforces
compliance with coming deprecation rules in PHP 8.4 concerning nullable
types. If the code does not remove the soon-to-be-deprecated form of
nullable type declarations, the library will no longer be usable in
Mediawiki for PHP 8.4 deployments.

Replace existing phpcs rules with the standard mediawiki-codesniffer
ruleset.

Bug: T379481